smooth and clip the image as necessary
smoothing ensures it looks dapper. ideally we'd want mipmapping but there
is a performance cost associated with it that may not be practical. in fact
loading the view already takes too long I find. possibly a loder might help
and/or image asyncness

clip ensures we dont paint outside the bounding rect. this isn't a problem
with the current fill mode, but could become one if we change the mode.